requirements.txt
m3u8==6.0.0
Requests==2.32.3
python-dotenv==1.0.1
